coverity defect fixes
