+++ archive/src/main/java/opendap/tools/GeturlApplet.java
@@ -1,0 +1,87 @@
+/////////////////////////////////////////////////////////////////////////////
+// This file is part of the "Java-DAP" project, a Java implementation
+// of the OPeNDAP Data Access Protocol.
+//
+// Copyright (c) 2010, OPeNDAP, Inc.
+// Copyright (c) 2002,2003 OPeNDAP, Inc.
+// 
+// Author: James Gallagher <jgallagher@opendap.org>
+// 
+// All rights reserved.
+// 
+// Redistribution and use in source and binary forms,
+// with or without modification, are permitted provided
+// that the following conditions are met:
+// 
+// - Redistributions of source code must retain the above copyright
+//   notice, this list of conditions and the following disclaimer.
+// 
+// - Redistributions in binary form must reproduce the above copyright
+//   notice, this list of conditions and the following disclaimer in the
+//   documentation and/or other materials provided with the distribution.
+// 
+// - Neither the name of the OPeNDAP nor the names of its contributors may
+//   be used to endorse or promote products derived from this software
+//   without specific prior written permission.
+// 
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+// IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+// TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+// PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+// HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+// SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
+// TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+// PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+// LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+// NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+/////////////////////////////////////////////////////////////////////////////
+
+
+
+package opendap.tools;
+
+import opendap.tools.GeturlFrame;
+
+import java.applet.*;
+import java.awt.*;
+import java.awt.event.*;
+
+/**
+ * Java geturl applet.
+ */
+public class GeturlApplet extends Applet {
+    /**
+     * The currently open Geturl window.
+     */
+    protected GeturlFrame frame;
+
+    /**
+     * Open the frame and add the start button to the layout
+     */
+    public void init() {
+        frame = new GeturlFrame(true);
+        Button restartButton = new Button("Restart");
+        restartButton.addActionListener(new ActionListener() {
+            public void actionPerformed(ActionEvent e) {
+                frame.dispose();
+                frame = new GeturlFrame(true);
+            }
+        });
+        add(restartButton);
+    }
+
+    /**
+     * Dispose of the frame
+     */
+    public void dispose() {
+        frame.dispose();
+    }
+
+    /** Main function to call as an application. */
+    public static void main(String args[]) {
+        new GeturlFrame(false);
+    }
+}
+
+
