+++ archive/src/main/java/opendap/tools/geturl
@@ -1,0 +1,23 @@
+#!/bin/sh
+#
+# $Id: geturl 15901 2007-02-28 23:57:28Z jimg $
+# Run the Java geturl client. A smarter script would figure out where the
+# code was installed and add that to CLASSPATH. 4/2/2002 jhrg
+
+CLASSPATH=$CLASSPATH:/home/jimg/Java-OPeNDAP/build/classes
+export CLASSPATH
+
+java opendap/util/geturl/Geturl "$@"
+
+# $Log: geturl,v $
+# Revision 1.1  2003/08/12 23:51:27  ndp
+# Mass check in to begin Java-OPeNDAP development work
+#
+# Revision 1.1  2002/04/03 12:36:08  ndp
+# Moved jgal's geturl shell script to the right place
+#
+# Revision 1.1  2002/04/02 17:28:40  jimg
+# Created. Needs to be hooked up with some sort of smart installer. For now
+# this just keeps me from fumbling whenever I try to run this particular Java
+# client...
+#
