+++ archive/src/main/java/opendap/tools/geturl.html
@@ -1,0 +1,10 @@
+<HTML>
+<HEAD>
+    <TITLE>Geturl</TITLE>
+</HEAD>
+
+<BODY>
+<APPLET CODE="opendap/client/gui/GeturlApplet.class" ARCHIVE="geturl.jar" WIDTH=80 HEIGHT=25>
+</APPLET>
+</BODY>
+</HTML>
