+++ cdm/src/main/java/ucar/nc2/ft/point/remote/RemotePointCollection.java
@@ -33,6 +33,7 @@
 
 package ucar.nc2.ft.point.remote;
 
+import com.sun.istack.internal.NotNull;
 import ucar.nc2.ft.point.PointCollectionImpl;
 import ucar.nc2.ft.point.PointIteratorEmpty;
 import ucar.nc2.ft.PointFeatureIterator;
@@ -73,7 +74,7 @@ class RemotePointCollection extends PointCollectionImpl implements QueryMaker {
   @Override
   public PointFeatureIterator getPointFeatureIterator(int bufferSize) throws IOException {
     InputStream in = null;
-    String errMessage = null;
+    String errMessage;
 
     try {
       in = CdmRemote.sendQuery(null, uri, queryMaker.makeQuery());
@@ -117,6 +118,7 @@ class RemotePointCollection extends PointCollectionImpl implements QueryMaker {
   // Must override default subsetting implementation for efficiency
 
   @Override
+  @NotNull
   public PointFeatureCollection subset(LatLonRect boundingBox, CalendarDateRange dateRange) throws IOException {
     return new PointFeatureCollectionSubset(this, boundingBox, dateRange);
   }
