--- opendap/src/main/java/opendap/tools/gui/GuiTest.java
@@ -1,82 +1,0 @@
-/////////////////////////////////////////////////////////////////////////////
-// This file is part of the "Java-DAP" project, a Java implementation
-// of the OPeNDAP Data Access Protocol.
-//
-// Copyright (c) 2010, OPeNDAP, Inc.
-// Copyright (c) 2002,2003 OPeNDAP, Inc.
-// 
-// Author: James Gallagher <jgallagher@opendap.org>
-// 
-// All rights reserved.
-// 
-// Redistribution and use in source and binary forms,
-// with or without modification, are permitted provided
-// that the following conditions are met:
-// 
-// - Redistributions of source code must retain the above copyright
-//   notice, this list of conditions and the following disclaimer.
-// 
-// - Redistributions in binary form must reproduce the above copyright
-//   notice, this list of conditions and the following disclaimer in the
-//   documentation and/or other materials provided with the distribution.
-// 
-// - Neither the name of the OPeNDAP nor the names of its contributors may
-//   be used to endorse or promote products derived from this software
-//   without specific prior written permission.
-// 
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
-// IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
-// TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
-// PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-// HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-// SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
-// TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
-// PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
-// LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
-// NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-/////////////////////////////////////////////////////////////////////////////
-
-
-package opendap.tools.gui;
-
-import opendap.tools.gui.msg_box;
-import opendap.tools.gui.warning_box;
-import opendap.tools.gui.error_box;
-
-/**
- * Test routine for the opendap/util/gui classes
- *
- * @author ndp
- * @version $Revision: 15901 $
- */
-
-public class GuiTest {
-
-
-    // Constructor
-    public GuiTest() {
-    }
-
-
-    public static void main(String[] args) {
-
-        msg_box mbox = new msg_box("Wow! A Message Box!");
-        msg_box mbox1 = new msg_box("My Title1",
-                "Wow! A Message Box! Maybe I should test a really long string in it to see if it automatically resizes for the text contained within it?\nOk!\nLets do That!");
-
-        warning_box wbox = new warning_box("Wow! A Warning Message Box!");
-        warning_box wbox1 = new warning_box("My Title2",
-                "Wow! A Warning Message Box!");
-
-        new error_box("Wow! A Error Message Box!");
-        new error_box("My Title3","Wow! A Error Message Box!");
-
-        System.exit(0);
-    }
-
-
-}
-
-
-
