--- opendap/src/main/java/opendap/tools/gui/error_box.java
@@ -1,87 +1,0 @@
-/////////////////////////////////////////////////////////////////////////////
-// This file is part of the "Java-DAP" project, a Java implementation
-// of the OPeNDAP Data Access Protocol.
-//
-// Copyright (c) 2010, OPeNDAP, Inc.
-// Copyright (c) 2002,2003 OPeNDAP, Inc.
-// 
-// Author: James Gallagher <jgallagher@opendap.org>
-// 
-// All rights reserved.
-// 
-// Redistribution and use in source and binary forms,
-// with or without modification, are permitted provided
-// that the following conditions are met:
-// 
-// - Redistributions of source code must retain the above copyright
-//   notice, this list of conditions and the following disclaimer.
-// 
-// - Redistributions in binary form must reproduce the above copyright
-//   notice, this list of conditions and the following disclaimer in the
-//   documentation and/or other materials provided with the distribution.
-// 
-// - Neither the name of the OPeNDAP nor the names of its contributors may
-//   be used to endorse or promote products derived from this software
-//   without specific prior written permission.
-// 
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
-// IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
-// TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
-// PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-// HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-// SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
-// TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
-// PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
-// LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
-// NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-/////////////////////////////////////////////////////////////////////////////
-
-
-
-package opendap.tools.gui;
-
-import javax.swing.*;
-
-/**
- * Dispays a simple error message box.
- *
- * @author ndp
- * @version $Revision: 15901 $
- */
-
-public class error_box {
-
-
-    /**
-     * Construct and display an error dialog box.
-     *
-     * @param msg This string is used as the message text in the dialog box.
-     */
-    public error_box(String msg) {
-        JLabel jl = new JLabel(msg);
-        JOptionPane.showMessageDialog(null,    // parent frame
-                jl,    // Object to Display
-                "ERROR!", // Title bar label
-                JOptionPane.ERROR_MESSAGE);
-    }
-
-
-    /**
-     * Construct and display an error dialog box.
-     *
-     * @param title This String is used in the title bar of the dialog box.
-     * @param msg   This string is used as the message text in the dialog box.
-     */
-    public error_box(String title, String msg) {
-        JLabel jl = new JLabel(msg);
-        JOptionPane.showMessageDialog(null,    // parent frame
-                jl,    // Object to Display
-                title, // Title bar label
-                JOptionPane.ERROR_MESSAGE);
-    }
-
-}
-
-
-
