--- opendap/src/main/java/opendap/tools/package.html
@@ -1,11 +1,0 @@
-<HTML>
-<HEAD>
-    <TITLE>Classes for Geturl, the sample OPeNDAP DAP2 client.</TITLE>
-</HEAD>
-
-<BODY>
-This package contains classes to implement Geturl, the sample OPeNDAP DAP2 client
-the GUI status window class used by Geturl.
-client.
-</BODY>
-</HTML>
