+++ opendap/src/main/java/ucar/nc2/dods/DodsV.java
@@ -585,20 +585,21 @@ class DodsV implements Comparable {
   ////////////////////////////////////////////////////////////////////////////////
   private static void doit(String urlName) throws IOException, DAP2Exception {
     System.out.println("DODSV read =" + urlName);
-    DConnect2 dodsConnection = new DConnect2(urlName, true);
+    try (DConnect2 dodsConnection = new DConnect2(urlName, true)) {
 
-    // get the DDS
-    DDS dds = dodsConnection.getDDS();
-    dds.print(System.out);
-    DodsV root = DodsV.parseDDS(dds);
+      // get the DDS
+      DDS dds = dodsConnection.getDDS();
+      dds.print(System.out);
+      DodsV root = DodsV.parseDDS(dds);
 
-    // get the DAS
-    DAS das = dodsConnection.getDAS();
-    das.print(System.out);
-    root.parseDAS(das);
+      // get the DAS
+      DAS das = dodsConnection.getDAS();
+      das.print(System.out);
+      root.parseDAS(das);
 
-    // show the dodsV tree
-    root.show(System.out, "");
+      // show the dodsV tree
+      root.show(System.out, "");
+    }
   }
 
   public static void main(String args[]) throws IOException, DAP2Exception {
