+++ tds/src/main/java/thredds/featurecollection/InvDatasetFeatureCollection.java
@@ -69,9 +69,9 @@ import java.util.*;
 /**
  * Abstract superclass for Feature Collection Datasets.
  * This is a InvCatalogRef subclass. So the reference is placed in the parent, but
- *  the catalog itself isnt constructed until the following call from DataRootHandler.makeDynamicCatalog():
- *   match.dataRoot.featCollection.makeCatalog(match.remaining, path, baseURI);
- * <p/>
+ * the catalog itself isnt constructed until the following call from DataRootHandler.makeDynamicCatalog():
+ * match.dataRoot.featCollection.makeCatalog(match.remaining, path, baseURI);
+ * <p>
  * The DatasetFeatureCollection object is held in the DataRootManager's FeatureCollectionCache; it may get closed and recreated.
  *
  * @author caron
@@ -88,6 +88,7 @@ public abstract class InvDatasetFeatureCollection {
   // can be changed
   static protected AllowedServices allowedServices;
   static protected String contextName = "/thredds";  // set by TdsInit
+
   static public void setLoggerFactory(LoggerFactory fac) {
     loggerFactory = fac;
   }
@@ -96,6 +97,7 @@ public abstract class InvDatasetFeatureCollection {
   static public void setContextName(String c) {
     contextName = c;
   }
+
   static public void setAllowedServices(AllowedServices _allowedServices) {
     allowedServices = _allowedServices;
   }
@@ -190,25 +192,26 @@ public abstract class InvDatasetFeatureCollection {
     orgService = parent.getServiceDefault();
     if (orgService == null) {
       String orgServiceName = parent.getServiceNameDefault();
-      if (orgServiceName != null)
-        orgService = allowedServices.findGlobalService(orgServiceName);
+      orgService = allowedServices.findGlobalService(orgServiceName);
     }
     if (orgService == null) {
-      orgService = allowedServices.getStandardServices( fcType.getFeatureType());
+      orgService = allowedServices.getStandardServices(fcType.getFeatureType());
     }
+    if (orgService == null)
+      return;
 
     if (orgService.getType() != ServiceType.Compound) {
       virtualService = orgService;
       return;
     }
 
+    // remove http service
     List<Service> nestedOk = new ArrayList<>();
     for (Service service : orgService.getNestedServices()) {
       if (service.getType() != ServiceType.HTTPServer) {
         nestedOk.add(service);
       }
     }
-
     virtualService = new Service("VirtualServices", "", ServiceType.Compound.toString(), null, null, nestedOk, orgService.getProperties());
   }
 
@@ -227,7 +230,7 @@ public abstract class InvDatasetFeatureCollection {
   //////////////////////////////////////////////////////
   // called by eventBus
   @Subscribe
-  public void processEvent(CollectionUpdateEvent event)  {
+  public void processEvent(CollectionUpdateEvent event) {
     if (!config.collectionName.equals(event.getCollectionName())) return; // not for me
 
     try {
@@ -357,7 +360,7 @@ public abstract class InvDatasetFeatureCollection {
     if (config.gribConfig.latestNamer != null) {
       return config.gribConfig.latestNamer;
     } else {
-      return "Latest "+name+" File";
+      return "Latest " + name + " File";
     }
   }
 
@@ -381,7 +384,7 @@ public abstract class InvDatasetFeatureCollection {
 
   protected String makeFullName(DatasetNode ds) {
     if (ds.getParent() == null) return ds.getName();
-    String parentName = makeFullName( ds.getParent());
+    String parentName = makeFullName(ds.getParent());
     if (parentName == null || parentName.length() == 0) return ds.getName();
     return parentName + "/" + ds.getName();
   }
@@ -429,7 +432,7 @@ public abstract class InvDatasetFeatureCollection {
     Catalog parentCatalog = parent.getParentCatalog();
 
     CatalogBuilder result = new CatalogBuilder();
-    result.setName( makeFullName(parent));
+    result.setName(makeFullName(parent));
     result.setVersion(parentCatalog.getVersion());
     result.setBaseURI(catURI);
     result.addService(orgService);
@@ -498,7 +501,7 @@ public abstract class InvDatasetFeatureCollection {
       String mapUri = URLnaming.resolve(baseURI.toString(), href);
       return new ThreddsMetadata.UriResolved(href, new URI(mapUri));
     } catch (Exception e) {
-      logger.error(" ** Invalid URI= '"+baseURI.toString()+"' href='"+href+"'%n", e);
+      logger.error(" ** Invalid URI= '" + baseURI.toString() + "' href='" + href + "'%n", e);
       return null;
     }
   }
@@ -510,7 +513,7 @@ public abstract class InvDatasetFeatureCollection {
     Catalog parentCatalog = parent.getParentCatalog();
 
     CatalogBuilder result = new CatalogBuilder();
-    result.setName( makeFullName(parent));
+    result.setName(makeFullName(parent));
     result.setVersion(parentCatalog.getVersion());
     result.setBaseURI(catURI);
     result.addService(orgService);
@@ -531,8 +534,8 @@ public abstract class InvDatasetFeatureCollection {
     String fname = mpath.substring(topDirectory.length() + 1);
 
     String path = FILES + "/" + fname;
-    top.put( Dataset.UrlPath, this.configPath + "/" + path);
-    top.put( Dataset.Id, this.configPath + "/" + path);
+    top.put(Dataset.UrlPath, this.configPath + "/" + path);
+    top.put(Dataset.Id, this.configPath + "/" + path);
     String lpath = this.configPath + "/" + path;
     top.put(Dataset.VariableMapLinkURI, new ThreddsMetadata.UriResolved(makeMetadataLink(lpath, VARIABLES), catURI));
     top.put(Dataset.DataSize, mfile.getLength());
