+++ tds/src/main/java/thredds/server/ncss/view/dsg/point/AbstractPointSubsetWriter.java
@@ -44,48 +44,44 @@ import java.util.List;
  * Created by cwardgar on 2014/06/02.
  */
 public abstract class AbstractPointSubsetWriter extends AbstractDsgSubsetWriter {
-    protected final PointFeatureCollection pointFeatureCollection;
-    private boolean headerDone;
+  protected final PointFeatureCollection pointFeatureCollection;
 
-    public AbstractPointSubsetWriter(FeatureDatasetPoint fdPoint, NcssParamsBean ncssParams)
-            throws NcssException, IOException {
-        super(fdPoint, ncssParams);
+  public AbstractPointSubsetWriter(FeatureDatasetPoint fdPoint, NcssParamsBean ncssParams)
+          throws NcssException, IOException {
+    super(fdPoint, ncssParams);
 
-        List<FeatureCollection> featColList = fdPoint.getPointFeatureCollectionList();
-        assert featColList.size() == 1 : "Is there ever a case when this is NOT 1?";
-        assert featColList.get(0) instanceof PointFeatureCollection :
-                "This class only deals with PointFeatureCollections.";
+    List<FeatureCollection> featColList = fdPoint.getPointFeatureCollectionList();
+    assert featColList.size() == 1 : "Is there ever a case when this is NOT 1?";
+    assert featColList.get(0) instanceof PointFeatureCollection :
+            "This class only deals with PointFeatureCollections.";
 
-        this.pointFeatureCollection = (PointFeatureCollection) featColList.get(0);
+    this.pointFeatureCollection = (PointFeatureCollection) featColList.get(0);
+  }
+
+  public abstract void writeHeader(PointFeature pf) throws Exception;
+
+  public abstract void writePoint(PointFeature pointFeat) throws Exception;
+
+  public abstract void writeFooter() throws Exception;
+
+  @Override
+  public void write() throws Exception {
+
+    // Perform spatial and temporal subset.
+    PointFeatureCollection subsettedPointFeatColl =
+            pointFeatureCollection.subset(ncssParams.getLatLonBoundingBox(), wantedRange);
+    if (subsettedPointFeatColl == null) // means theres nothing in the subset
+      return;
+
+    boolean headerDone = false;
+    for (PointFeature pointFeat : subsettedPointFeatColl) {
+      if (!headerDone) {
+        writeHeader(pointFeat);
+        headerDone = true;
+      }
+      writePoint(pointFeat);
     }
 
-    public abstract void writeHeader(PointFeature pf) throws Exception;
-
-    public abstract void writePoint(PointFeature pointFeat) throws Exception;
-
-    public abstract void writeFooter() throws Exception;
-
-    @Override
-    public void write() throws Exception {
-
-        // Perform spatial and temporal subset.
-        PointFeatureCollection subsettedPointFeatColl =
-                pointFeatureCollection.subset(ncssParams.getLatLonBoundingBox(), wantedRange);
-
-        subsettedPointFeatColl.resetIteration();
-        try {
-            while (subsettedPointFeatColl.hasNext()) {
-                PointFeature pointFeat = subsettedPointFeatColl.next();
-                if (!headerDone) {
-                  writeHeader(pointFeat);
-                  headerDone = true;
-                }
-                writePoint(pointFeat);
-            }
-        } finally {
-            subsettedPointFeatColl.finish();
-        }
-
-        writeFooter();
-    }
+    writeFooter();
+  }
 }
